UIColor只保留最常用的

+++ FYHelper/FYCategory/UIColor+Colours.m
@@ -27,7 +27,7 @@
 #pragma mark - Static Block
 
 static CGFloat (^RAD)(CGFloat) = ^CGFloat(CGFloat degree) {
-    return degree * M_PI / 180;
+    return (CGFloat) (degree * M_PI / 180);
 };
